# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Gacha Game Backend API
  version: 1.0.0
  description: API for managing gacha rolls, user inventories, auctions, transactions, and user authentication.
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /gacha/add:
    post:
      tags:
        - GACHA
      summary: Add a new gacha item
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                rarity:
                  type: string
                status:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Gacha item added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing data to add gacha item
        '500':
          description: Database error
  /gacha/roll:
    post:
      tags:
        - GACHA
      summary: Perform a gacha roll for a random item
      responses:
        '200':
          description: Gacha roll successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  gacha_id:
                    type: string
                  name:
                    type: string
                  rarity:
                    type: string
        '400':
          description: Missing data for gacha roll
        '403':
          description: Insufficient funds for gacha roll
        '500':
          description: Failed to add gacha item to user inventory
  /gacha/inventory/{user_id}:
    get:
      tags:
        - GACHA
      summary: Retrieve a user's gacha inventory
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User inventory retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventory:
                    type: array
                    items:
                      type: object
                      properties:
                        gacha_id:
                          type: string
                        name:
                          type: string
                        rarity:
                          type: string
                        status:
                          type: string
                        description:
                          type: string
                        acquired_date:
                          type: string
                        locked:
                          type: boolean
                        image:
                          type: string
        '404':
          description: No gacha items found for user
  /gacha/inventory/add:
    post:
      tags:
        - GACHA
      summary: Add a gacha item to a user's inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gacha_id:
                  type: string
      responses:
        '201':
          description: Gacha item successfully added to user's inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing data to add gacha to inventory
        '404':
          description: User or gacha item not found
  /gacha/all:
    get:
      tags:
        - GACHA
      summary: Retrieve all gacha items
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip
      responses:
        '200':
          description: Gacha items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: object
                      properties:
                        gacha_id:
                          type: string
                        name:
                          type: string
                        rarity:
                          type: string
                        status:
                          type: string
                        description:
                          type: string
                        image:
                          type: string
        '404':
          description: No gacha items found
  /gacha/update:
    put:
      tags:
        - GACHA
      summary: Update a gacha item
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                gacha_id:
                  type: string
                name:
                  type: string
                rarity:
                  type: string
                status:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Gacha item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing data to update gacha item

        '404':
          description: Gacha item not found
        '500':
          description: Failed to update gacha item
  /gacha/get/{gacha_id}:
    get:
      tags:
        - GACHA
      summary: Retrieve information about a specific gacha item
      parameters:
        - in: path
          name: gacha_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gacha item retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  gacha_id:
                    type: string
                  name:
                    type: string
                  rarity:
                    type: string
                  status:
                    type: string
                  description:
                    type: string
                  image:
                    type: string
        '404':
          description: Gacha item not found
  /gacha/get/{user_id}/{gacha_id}:
    get:
      tags:
        - GACHA
      summary: Retrieve details of a specific gacha item for a specific user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: gacha_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gacha item retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  gacha_id:
                    type: string
                  name:
                    type: string
                  rarity:
                    type: string
                  status:
                    type: string
                  description:
                    type: string
                  acquired_date:
                    type: string
                  locked:
                    type: boolean
                  image:
                    type: string
        '404':
          description: User not found or gacha item not found in user inventory
  /gacha/is_gacha_unlocked/{user_id}/{gacha_id}:
    get:
      tags:
        - GACHA
      summary: Check if a gacha item is unlocked for a user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: gacha_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gacha item is unlocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: Gacha item is locked
        '400':
          description: Missing data to check gacha item
  /gacha/update_gacha_status:
    put:
      tags:
        - GACHA
      summary: Update the status of a gacha item for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                gacha_id:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Gacha status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing data to update gacha status
        '409':
          description: User not found
        '408':
          description: Gacha item not found
        '500':
          description: Failed to update gacha status
  /gacha/update_gacha_owner:
    put:
      tags:
        - GACHA
      summary: Update the owner of a gacha item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                buyer_id:
                  type: string
                seller_id:
                  type: string
                gacha_id:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Gacha owner updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing data to update gacha owner
        '404':
          description: Buyer not found or gacha item not found
        '500':
          description: Failed to update gacha owner
  /gacha/delete/{gacha_id}:
    delete:
      tags:
      - GACHA
      summary: Delete a gacha item from the system (admin only)
      parameters:
        - name: gacha_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gacha item successfully deleted
        '404':
          description: Gacha not found
        '500':
          description: Failed to delete gacha item